Refactor GATT server
