+++ src/features/ble/gatt-server/available-wifi-characteristics.ts
