+++ src/features/ble/gatt-server/current-wifi-characteristics.ts
