+++ src/features/ble/gatt-server/device-info-characteristics.ts
